fpxpy/pyproject.toml
[project]
name = "fpxpy"
version = "0.3.3"
authors = [
    { name = "Fiberplane", email = "support@fiberplane.com" },
    { name = "mellowagain", email = "py@mari.zip" },
    { name = "flenter", email = "flenter@users.noreply.github.com" }
]
description = "Fiberplane studio integration with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dataclasses-json>=0.6.7",
    "fastapi>=0.108.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-sdk>=1.29.0",
    "requests>=2.32.3",
    "tracing>=0.1.3",
    "types-requests>=2.32.0.20241016",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/fiberplane/fpx"
Issues = "https://github.com/fiberplane/fpx/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "ruff>=0.8.6",
]
